draft for internal interface
